# openapi/esim_go_schema_v2_5.yaml
openapi: 3.0.1
info:
  title: eSIM-Go API
  version: 2.5.0
  description: |
    To access the eSIMGo API, you need to authenticate your requests using an API key.
    Here's a brief overview of the authentication process:
      1. Assuming that you have created account already:
      - Log into your eSIMGo account at https://sso.esim-go.com/login.
      - Navigate to Account Settings -> API Details to find your API key.
      - Keep your API key secure and avoid sharing it with others.
      2. API Key Usage:
      - Include your API key in the header of all eSIMGo API requests.
      - Use the header key 'X-API-KEY' with your API key as the value.
      3. Security Scheme:
      - The eSIMGo API uses an HTTP security scheme type for authentication.
      4. Authorization Types:
      - The API supports 'apiKeyAuth' authorization type.
servers:
  - url: https://api.esim-go.com/v2.5
    description: eSIMGo API
...
paths:
...
  /esims/qr/{reference}:
    get:
      summary: Get QR codes for eSIMs from an Order or Bundle apply reference (Deprecated)
      description: |
        Get multiple QR codes by providing an Order or Apply Reference. Returns multiple QR codes as `.PNG` contained in a `.ZIP`.

        Deactivated eSIMs will return a 410 Gone error.

        Deprecated: See `/esims/assignments`
      tags:
        - eSIMs (Deprecated)
      parameters:
        - in: path
          name: reference
          schema:
            type: string
          required: true
          description: Order Reference or Apply Reference
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: QR code image in a ZIP file
          content:
            application/zip:
              schema:
                type: zip
                example: cfb94100-1c41-4d0e-9ff2-0dsfn23vd1121233cda46.zip
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid or the user does not have permission to query the given ICCID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "410":
          description: Gone - returned when the eSIM has been deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -g 'https://api.esim-go.com/v2.5/esims/qr/{reference}' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims/qr/{reference}',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims/qr/{reference}"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims/qr/{reference}"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/esims/qr/{reference}");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/esims/qr/{reference}")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
  /esims/csv/{reference}:
    get:
      summary: Get CSV containing eSIMs from an Order or Bundle apply (Deprecated)
      description: |
        Get eSIM SMDP+ details by providing an Order or Apply Reference.

        Contains ICCID, SMDP+ address and Matching ID.

        Deprecated: See `/esims/assignments`
      tags:
        - eSIMs (Deprecated)
      parameters:
        - in: path
          name: reference
          schema:
            type: string
          required: true
          description: Order Reference or Apply Reference
        - in: query
          name: additionalFields
          schema:
            type: array
            items:
              type: string
              enum:
                - appleInstallUrl
                - installUrl
          explode: true
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: CSV
          content:
            text/csv:
              schema:
                type: string
                example: |
                  "ICCID","Matching ID","RSP URL","Bundle"
                  "Value 1","Value 2","Value 3","Value 4"
        "400":
          description: Bad Request - returned when the given ICCID is not found and/or permission to view this ICCID is not granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -g 'https://api.esim-go.com/v2.5/esims/csv/{reference}' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims/csv/{reference}',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims/csv/{reference}"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims/csv/{reference}"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/esims/csv/{reference}");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/esims/csv/{reference}")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
...
  /esims/{iccid}/bundles:
    get:
      summary: List Bundles applied to eSIM
      description: |
        This endpoint allows you to retrieve a list of all bundles that have been applied to a specific eSIM. This endpoint is useful for tracking the service history and current status of an eSIM.

        Remaining data can be found here.

        Each Bundle can have multiple assignments.

        Bundle Assignment States:
        - **Processing:**
          The bundle assignment is currently processing.
          This is usually instant but can, on occasion, take up to 10 minutes to complete.
          The eSIM can still be installed and will register on a network while the bundle is processing.
        - **Queued:**
          The bundle has been successfully assigned, has not been used yet, and is queued for use.
        - **Active:**
          The bundle has successfully been used.
          It has data remaining and is within the bundle duration.
        - **Depleted:**
          The bundle has no data remaining but is still within the bundle duration.
        - **Expired:**
          The bundle has expired, and the bundle duration has been exceeded.
        - **Revoked:**
          The bundle has been revoked, and is no longer on the esim.
        - **Lapsed:**
          The bundle has expired without being used and is no longer on the eSIM.
      tags:
        - eSIMs
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: includeUsed
          in: query
          schema:
            type: string
          description: |
            Include used & expired Bundles
            Backward compatibility for v2.1
          example: true
        - name: limit
          in: query
          schema:
            type: string
          description: |
            Number of assignments to return. Must be between 1 and 200. Default is 15
          example: 10
        - name: iccid
          in: path
          schema:
            type: string
          required: true
          description: (Required) The ICCID of the eSIM
          example: 8943108165016100000
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundlesList"
              example:
                bundles:
                  - name: esim_1GB_7D_GB_V2
                    description: eSIM, 1GB, 7 Days, United Kingdom, V2
                    assignments:
                      - id: "215009266"
                        callTypeGroup: data
                        initialQuantity: 1000000000
                        remainingQuantity: 1000000000
                        assignmentDateTime: "2024-12-11T10:47:00.34523Z"
                        assignmentReference: 3a6e2e40-c674-45e9-b289-b35027bf72a9-0
                        bundleState: queued
                        unlimited: false
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -g 'https://api.esim-go.com/v2.5/esims/{iccid}/bundles' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims/{iccid}/bundles',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims/{iccid}/bundles"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims/{iccid}/bundles"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/esims/{iccid}/bundles");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/esims/{iccid}/bundles")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
    post:
      summary: Apply a Bundle to an eSIM (Deprecated)
      tags:
        - eSIMs (Deprecated)
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: iccid
          in: path
          schema:
            type: string
          required: true
          description: (Required) The ICCID of the eSIM
          example: <string>
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: <string>
                startTime: <string>
                repeat: <integer>
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                status: <string>
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -g 'https://api.esim-go.com/v2.5/esims/{iccid}/bundles' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json' \
            -d '{"type":"object","example":{"name":"<string>","startTime":"<string>","repeat":"<integer>"}}'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');
            let data = JSON.stringify({
              "type": "object",
              "example": {
                "name": "<string>",
                "startTime": "<string>",
                "repeat": "<integer>"
              }
            });

            let config = {
              method: 'post',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims/{iccid}/bundles',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              },
              data : data
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "strings"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims/{iccid}/bundles"
              method := "POST"

              payload := strings.NewReader(`{"type":"object","example":{"name":"<string>","startTime":"<string>","repeat":"<integer>"}}`)

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, payload)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims/{iccid}/bundles"

            payload = json.dumps({
              "type": "object",
              "example": {
                "name": "<string>",
                "startTime": "<string>",
                "repeat": "<integer>"
              }
            })
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("POST", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Post, "https://api.esim-go.com/v2.5/esims/{iccid}/bundles");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent("{\"type\":\"object\",\"example\":{\"name\":\"<string>\",\"startTime\":\"<string>\",\"repeat\":\"<integer>\"}}", null, "application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.post("https://api.esim-go.com/v2.5/esims/{iccid}/bundles")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .body("{\"type\":\"object\",\"example\":{\"name\":\"<string>\",\"startTime\":\"<string>\",\"repeat\":\"<integer>\"}}")
              .asString();
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      x-displayName: API Key in Header
      description: |
        Your API Key can passed to the eSIMGo API by setting a Header with the key `X-API-Key` in your HTTP call.

        e.g.

        ```
        curl -H 'X-API-Key: $API_KEY' https://api.esim-go.com/v2.3/esims/$ICCID/bundles
        ```
  schemas:
...
    Message:
      type: object
      properties:
        message:
          description: status of api functions which do not return data
          type: string
...
    BundlesList:
      type: object
      properties:
        bundles:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: Name of the bundle
              description:
                type: string
                description: Description of the bundle (size, validity period, coverage)
              assignments:
                type: array
                items:
                  type: object
                  properties:
                    allowances:
                      description: The allowances of the bundle
                      type: array
                      items:
                        type: object
                        properties:
                          type:
                            description: The type of allowance
                            type: string
                            enum:
                              - DATA
                              - VOICE
                              - SMS
                          service:
                            description: The service of the allowance, this defines where the allowance can be used
                            type: string
                            enum:
                              - STANDARD
                              - ROAMING
                          description:
                            description: The description of the allowance
                            type: string
                          initialAmount:
                            description: The initial amount of the allowance
                            type: integer
                            format: float
                          remainingAmount:
                            description: The remaining amount of the allowance
                            type: integer
                            format: float
                          unit:
                            description: The unit of the allowance
                            type: string
                            enum:
                              - MB
                              - SMS
                              - MINS
                          unlimited:
                            description: If the allowance is unlimited
                            type: boolean
                    id:
                      type: string
                      description: ID of assignment
                    callTypeGroup:
                      type: string
                      description: Type of the bundle
                    initialQuantity:
                      type: integer
                      description: The initial quantity the bundle had (in bytes)
                    remainingQuantity:
                      type: integer
                      description: The remaining quantity the bundle had (in bytes)
                    assignmentDateTime:
                      type: string
                      format: date-time
                      description: The date and time the bundle was created (utc string)
                    assignmentReference:
                      type: string
                      description: Assigment reference
                    bundleState:
                      type: string
                      description: Current state of a bundle
                    startTime:
                      type: string
                      format: date-time
                      description: The date and time the bundle was started (utc string)
                    endTime:
                      type: string
                      format: date-time
                      description: The date and time the bundle ends (utc string)
                    unlimited:
                      type: boolean
                      description: If the bundle is unlimited
...